ai_rules/requirements.txt
PyYAML>=6.0
requests>=2.32.0
python-dotenv>=1.0.1
google-api-python-client>=2.169.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
